TODODODO! Fix v10 solution configuration map
When the solution is in debug, all projects should be in debug. When
solution is in x64, all projects should be in x64. All projects with
appropriate targets are checked for inclusion in build. This might cause
some build errors from the IDE, but if the developers ever hope to bring
their own projects up to date, this will be future-proofing...

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Conflicts:
	MMCoreJ_wrap/MMCoreJ_wrap_v10.sln