TODODODO! Added build AcqEngine step & rebuild for plugins
The MM acqEngine also has to be built, and it doesn't seem to be built
elsewhere. Some plugins rely on it. mmstudio does also. I added a step to
compile it before the plugins.

I also added a quick bit to stop the plugins being cleaned unless
'--rebuild' was specified.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Conflicts:
	build-openspim.sh